Moved modal code. Created delete, info, tag modals.
This was a bit of a recoding/reworking of the original "drop downs".
Previously when adding tags, an animation expanded the view downward
with an input. This wasn't particularly useful so instead I created a
modal window which will be used to manage the tagging. Plus, delete and
info are also available modals. I figured it might be possible to break
the essense down for modals, thus I created a single modal obj that is
used as a base "class".

Signed-off-by: Nick Campbell <nicholas.j.campbell@gmail.com>